4.4 Component files updated
This commit updates the component files. An api.js file was created handling the requests to the meetup api.
In Event and EventList the data has been changed from state level to props.
Page and lat/lon elements added to interact with api and generate requests based on location and number of events requested